chore(localized-multiline-text-input): move to packages
