+++ packages/components/inputs/localized-multiline-text-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './localized-multiline-text-input';
+
+export { default as version } from './version';
