+++ packages/components/inputs/localized-multiline-text-input/src/localized-multiline-text-input.visualroute.js
@@ -5,7 +5,7 @@ import {
   ErrorMessage,
   WarningMessage,
 } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const lorem =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.';
