+++ packages/components/inputs/localized-multiline-text-input/src/translation-input.js
@@ -8,8 +8,8 @@ import { AngleUpIcon } from '@commercetools-uikit/icons';
 import Stack from '@commercetools-uikit/spacings-stack';
 import { filterDataAttributes } from '@commercetools-uikit/utils';
 import Text from '@commercetools-uikit/text';
-import MultilineInput from '../../internals/multiline-input';
-import messages from '../../internals/messages/multiline-input';
+import MultilineInput from '../../../../../src/components/internals/multiline-input';
+import messages from '../../../../../src/components/internals/messages/multiline-input';
 import {
   getTextareaStyles,
   getLanguageLabelStyles,
